add setTheme method
