+++ lib/plugins/dark_mode.mjs
@@ -7,6 +7,22 @@
 /**
 @function dark_mode
 */
+
+const darkTheme = {
+  base: '#222222',
+  'base-secondary': '#555555',
+  'base-tertiary': '#444444',
+  'base-contrast': '#333333',
+  'base-border': '#666666',
+  text: '#f2f2f2',
+  'text-secondary': '#CCCCCC',
+  'text-tertiary': '#ffffff',
+  accent: '#DAD095',
+  'accent-active': '#e18335',
+  blocked: '#ef5350',
+  changed: '#666600',
+};
+
 export function dark_mode() {
   // Get the map button
   const mapButton = document.getElementById('mapButton');
@@ -18,7 +34,7 @@ export function dark_mode() {
   let darkMode = localStorage.getItem('darkMode') === 'true';
 
   // toggle dark_mode if true.
-  darkMode && toggleDarkMode(darkMode);
+  darkMode && toggleDarkMode(darkTheme);
 
   // If the button container exists, append the dark mode button.
   mapButton.append(mapp.utils.html.node`<button
@@ -30,94 +46,11 @@ export function dark_mode() {
 
       localStorage.setItem('darkMode', darkMode);
 
-      toggleDarkMode(darkMode);
+      mapp.ui.utils.cssColourTheme(darkMode ? darkTheme : {});
 
       // change mode icon
       const modeIcon = mapButton.querySelector('.btn-color-mode span');
       modeIcon.textContent = darkMode ? 'light_mode' : 'dark_mode';
     }}>
     <span class="material-symbols-outlined">${darkMode ? 'light_mode' : 'dark_mode'}`);
-}
-
-/**
-Function to set the local storage to if the user uses darkmode or not.
-@function toggleDarkMode
-@description
-Colour palette consists of the following colours:
-* Backgrounds *
-base - most outer background, color used for document body, control panel.
-base-secondary - background colour for first level of containers. Might be the same as outer background for some light palettes.
-base-tertiary - background colour for interative elements embedded in the containers of the first level.
-base-contrast - colour used as baseline for palette that makes elements stand out. For light palettes this is expected to be white, for dark palettes - their darkest shade. Used as background for editable text etc.
-* Border * 
-base-border - colour used for borders. Might be the same as any of background colours or contrasting for better readability.
-* Text *
-text - default body text colour. 
-text-secondary - less prominent colour for helper text and symbols, ideally neutral colour readable on both darker and lighter backgrounds.
-text-tertiary - colour contrasting with default text, used on inverted backgrounds. Example: white text to use in label of accent colour.
-* Accents * 
-accent - prominent colour for interactive elements - buttons, links, toggles.
-accent-active - prominent colour for interactive elements in their activated state.
-accent-hover - colour used for highlighting elements on hover.
-* Information *
-info - prominent colour indicating information, confirmation, valid state. Background neutral green by default.
-blocked - prominent colour indicating error, invalid data, irreversible actions. Red by default.
-changed - colour used as background on edited fields. Subtle yellow by default, should fit well with text for readability.
-
-Defined colour variables will override corresponsing defaults.
-
-Palette template using default variables:
-```json 
-{
-      // Backgrounds
-      base: '#f2f2f2',
-      'base-secondary': '#f2f2f2',
-      'base-tertiary': '#fafafa',
-      'base-contrast': '#ffffff',
-      // Border
-      'base-border': '#dddddd',
-      // Text
-      text: '#3f3f3f',
-      'text-secondary': '#858585',
-      'text-tertiary': '#dddddd',
-      // Accents
-      accent: '#003d57',
-      'accent-active': '#d57120',
-      'accent-hover': '#939faa',
-      // Information
-      info: '#0b6f50',
-      blocked: '#a21309',
-      changed: '#ffffa7',
-}
-```
-*/
-function toggleDarkMode(darkMode) {
-  const root = document.querySelector(':root');
-
-  const dark_palette = {
-    base: '#222222',
-    'base-secondary': '#555555',
-    'base-tertiary': '#444444',
-    'base-contrast': '#333333',
-    'base-border': '#666666',
-    text: '#f2f2f2',
-    'text-secondary': '#CCCCCC',
-    'text-tertiary': '#ffffff',
-    accent: '#DAD095',
-    'accent-active': '#e18335',
-    blocked: '#ef5350',
-    changed: '#666600',
-  };
-
-  if (darkMode) {
-    // add defined variabled to the root style
-    Object.entries(dark_palette).map((entry) => {
-      root.style.setProperty(`--color-${entry[0]}`, `${entry[1]}`);
-    });
-  } else {
-    // remove variables from the root style. Bounce back to defaults in the ui stylesheet.
-    Object.entries(dark_palette).map((entry) => {
-      root.style.removeProperty(`--color-${entry[0]}`, `${entry[1]}`);
-    });
-  }
 }
