+++ lib/ui/utils/_utils.mjs
@@ -44,6 +44,8 @@ function Chart(_this) {
   );
 }
 
+import cssColourTheme from './cssColourTheme.mjs';
+
 import idleLogout from './idleLogout.mjs';
 
 import imagePreview from './imagePreview.mjs';
@@ -57,6 +59,7 @@ import locationCount from './locationCount.mjs';
 export default {
   Chart,
   Tabulator,
+  cssColourTheme,
   Json: jsonDataview,
   idleLogout,
   imagePreview,
